requirements.txt
chainlit>=0.7.0,<2.0.0
chromadb>=0.4.15
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
langchain-text-splitters>=0.0.1